split api & useradmin into their own routes
