+++ routes/index.js
@@ -1,0 +1,10 @@
+
+
+
+
+module.exports = (_models, _controllers, _logger) => {
+    return {
+        useradmin: require('./useradmin')(_models, _controllers, _logger),
+        api: require('./api')(_models, _controllers, _logger)
+    }
+}
